pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "alist-sync"
authors = [{ name = "LeeCQ", email = "lcq@leecq.cn" }]
description = "Alist多驱动同步工具."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alist-sdk",
    "typer",
    "rich",
    "async-lru",
    "pydantic",
    "pyyaml",
    "pymongo",
    "psutil",
]
dynamic = ["version"]

classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Framework :: Pydantic :: 2",
    "Natural Language :: Chinese (Simplified)",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/lee-cq/alist-sync"
Issues = "https://github.com/lee-cq/alist-sync/issues"

[tool.setuptools]
packages = ["alist_sync"]

[tool.setuptools.dynamic]
version = { attr = "alist_sync.version.__version__" }